python/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-dotenv==1.2.2
requests==2.33.0
aiohttp==3.13.3
httpx==0.28.1
pydantic==2.12.5
